Fix platform dependent compilation for OSX (#96)
On macOS, we want the code blocks take effect in the Editor, regardless of the chosen build target. So in addition to `UNITY_STANDALONE_OSX`, we also have to check for `UNITY_EDITOR_OSX`.

And on Windows, we don't want these code blocks to take effect in the Editor, even though macOS has been selected as the build target. So we only should check for `UNITY_STANDALONE_OSX` when not running in the Editor.